requirements.txt
appdirs==1.4.4
asgiref==3.4.0
backcall==0.2.0
backports.csv==1.0.7
certifi==2022.9.24
chardet==3.0.4
contextlib2==21.6.0
decorator==5.0.9
distlib==0.3.2
Django==3.2.4
django-appconf==1.0.5
django-extensions==3.1.3
django-imagekit==4.0.2
django-webpack-loader==0.6.0
djem==0.6.4
docopt==0.6.2
filelock==3.0.12
flake8==3.9.2
graphviz==0.20
gunicorn==20.1.0
idna==2.8
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mccabe==0.6.1
numpy==1.22.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==8.3.2
pipenv==2021.5.29
pp-ez==0.2.0
pprintpp==0.4.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
python-dotenv==0.18.0
pytz==2021.1
requests==2.22.0
schema==0.7.4
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.2
traitlets==5.0.5
urllib3==1.25.11
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
coverage==6.5.0
djangorestframework==3.14.0